+++ tfpy/main.py
@@ -1,21 +1,50 @@
 """Define the application entrypoint."""
 import importlib.util
 import json
+import logging
 from pathlib import Path
 import sys
 
-import typer
+from loguru import logger
 from terraformpy import TFObject
+import typer
 
 from tfpy.core.stack import StackVars
 
 
 app = typer.Typer()
 
+# Configure logger.
+INITIAL_LOG_LEVEL = logging.WARNING
+LOG_FORMAT_COMPACT = "<level>{message}</level>"
+LOG_FORMAT_VERBOSE = (
+    "<level>{time:YYYY-MM-DDTHH:mm:ssZZ} {name}:{line:<4} {message}</level>"
+)
 
+# Remove any predefined logger.
+logger.remove()
+
+# Set the log colors.
+logger.level("ERROR", color="<red><bold>")
+logger.level("WARNING", color="<yellow>")
+logger.level("SUCCESS", color="<green>")
+logger.level("INFO", color="<cyan>")
+logger.level("DEBUG", color="<blue>")
+logger.level("TRACE", color="<magenta>")
+
+# pylint: disable=C0330
 @app.command()
-def generate(project: str, environment: str = ""):
+def generate(
+    project: str,
+    environment: str = "",
+    verbose: int = typer.Option(0, "--verbose", "-v", count=True),
+):
     """Generate Terraform stacks."""
+    # Configure logger verbosity .
+    log_level = max(INITIAL_LOG_LEVEL - verbose * 10, 0)
+    log_format = LOG_FORMAT_VERBOSE if log_level < logging.INFO else LOG_FORMAT_COMPACT
+    logger.add(sys.stderr, format=log_format, level=log_level, colorize=True)
+
     # Load the stackvars.
     stackvars = StackVars(project, environment=environment, var_dir=Path.cwd(),)
     stackvars.load()
@@ -25,6 +54,7 @@ def generate(project: str, environment: str = ""):
     library = Path("library")
     for lib in library.glob("**/*.py"):
         # Import the libs.
+        logger.debug(f'Importing library "{lib}".')
         spec = importlib.util.spec_from_file_location(lib.stem, lib)
         module = importlib.util.module_from_spec(spec)
         sys.modules[lib.stem] = module
@@ -34,6 +64,7 @@ def generate(project: str, environment: str = ""):
     s = Path("stacks")
     for stack in s.glob("**/*.tf.py"):
         # Import the stack.
+        logger.debug(f'Importing stack "{stack}".')
         spec = importlib.util.spec_from_file_location(
             stack.name.replace("".join(stack.suffixes), ""), stack
         )