+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add logging support to the application. [#]
+
 ## [0.2.0] - 2020-04-26
 
 ### Added